To java big int and big dec.

+++ plugins/org.openmole.ide.plugin.prototype.base/src/main/scala/org/openmole/ide/plugin/prototype/base/BigIntegerPrototypeDataUI.scala
@@ -17,16 +17,17 @@
 
 package org.openmole.ide.plugin.prototype.base
 
+import java.math.BigInteger
 import org.openmole.core.implementation.data.Prototype
 import org.openmole.core.implementation.data.Prototype._
 import org.openmole.core.model.data.IPrototype
 
-class BigIntegerPrototypeDataUI(val name: String="", d: Int=0) extends GenericPrototypeDataUI[BigInt](d){
+class BigIntegerPrototypeDataUI(val name: String="", d: Int=0) extends GenericPrototypeDataUI[BigInteger](d){
   def displayTypedName = displayName + " : BigInt"
   
-  def coreClass = classOf[IPrototype[BigInt]]
+  def coreClass = classOf[IPrototype[BigInteger]]
   
-  def coreObject = new Prototype[BigInt](name).toArray(dim).asInstanceOf[IPrototype[BigInt]]
+  def coreObject = new Prototype[BigInteger](name).toArray(dim).asInstanceOf[IPrototype[BigInteger]]
   
   def fatImagePath = "img/biginteger_fat.png"
